+++ nvlib/nvImage/src/nvImage.cpp
@@ -87,7 +87,7 @@ const void* Image::getLevel( int level, GLenum face) const {
 
     face = face - GL_TEXTURE_CUBE_MAP_POSITIVE_X;
 
-    assert( (face*_levelCount + level) < (int)_data.size());
+    assert( (int)(face*_levelCount + level) < (int)_data.size());
     return _data[ face*_levelCount + level];
 }
 
@@ -100,7 +100,7 @@ void* Image::getLevel( int level, GLenum face) {
 
     face = face - GL_TEXTURE_CUBE_MAP_POSITIVE_X;
 
-    assert( (face*_levelCount + level) < (int)_data.size());
+    assert( (int)(face*_levelCount + level) < (int)_data.size());
     return _data[ face*_levelCount + level];
 }
